pyproject.toml
[tool.poetry]
name = "translation"
version = "0.1.0"
description = ""
authors = ["Camille-Bre <camille@may-sante.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.3"
openai = "^1.51.2"
sqlalchemy = "^2.0.35"
psycopg2-binary = "^2.9.9"
google-auth = "^2.35.0"
google-auth-oauthlib = "^1.2.1"
google-auth-httplib2 = "^0.2.0"
google-api-python-client = "^2.149.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
